+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symboltable/DeclarationFinderFunction.java
@@ -4,6 +4,7 @@
 
 package net.sourceforge.pmd.lang.java.symboltable;
 
+import net.sourceforge.pmd.lang.java.ast.ASTMethodReference;
 import net.sourceforge.pmd.lang.symboltable.NameDeclaration;
 import net.sourceforge.pmd.lang.symboltable.NameOccurrence;
 import net.sourceforge.pmd.util.SearchFunction;
@@ -19,6 +20,11 @@ public class DeclarationFinderFunction implements SearchFunction<NameDeclaration
 
     @Override
     public boolean applyTo(NameDeclaration nameDeclaration) {
+        // do not match method references
+        if (occurrence.getLocation() instanceof ASTMethodReference) {
+            return false;
+        }
+
         if (isDeclaredBefore(nameDeclaration) && isSameName(nameDeclaration)) {
             decl = nameDeclaration;
             return false;