+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/jdkversiontests/java10/LocalVariableTypeInferenceForLoopEnhanced2.java
@@ -1,0 +1,20 @@
+import java.util.ArrayList;
+import java.util.List;
+
+public class LocalVariableTypeInferenceForLoopEnhanced2 {
+    public void listMethod() {
+       List<String> test = new ArrayList<>();
+
+       for (var s : test) {
+            System.out.println(s);
+        }
+    }
+
+    public void arrayMethod() {
+        int values[] = {2, 4, 6};
+
+        for (var s : values) {
+             System.out.println(s);
+         }
+     }
+}